requirements.txt
streamlit>=1.35.0
requests>=2.31.0
ollama>=0.2.0
python-dotenv>=1.0.0


# File reading — multiple fallbacks, install all for best coverage
pdfplumber>=0.10.3        # Best PDF extraction
PyPDF2>=3.0.1             # PDF fallback #1
pypdf>=4.0.0              # PDF fallback #2 (newer fork)
pymupdf>=1.24.0           # PDF fallback #3 (fitz) — very fast

python-docx>=1.1.0        # DOCX reading
docx2txt>=0.8             # DOCX fallback

openpyxl>=3.1.2           # Excel reading
pandas>=2.0.0             # Excel/CSV fallback + data processing

# Optional extras
python-multipart>=0.0.9   # File upload support